Fix possible race condition
FileQueue#pop would lock the queue file and read the first line and
remainder of the file. It would then unlock, then relock to write the
remainder of the file.

During this brief unlocked phase, another thread could also attempt a
pop, read the first line and remainder, and then being the writing
phase. This would result in duplicate items being popped off the queue.

Instead, we're opening the file read-write, locking the file, and doing
all reading and modification while locked.